ena: Fix memory barrier comment
The dma_rmb() memory barrier guarantees that the device set the
phase bit before continuing to read the rest of the descriptor.
Because the phase bit and the rest of the descriptor are in the same
cache line this ensures coherency of the data from the descriptor.

Approved by: cperciva
MFC after: 2 weeks
Sponsored by: Amazon, Inc.